requirements.txt
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
python-telegram-bot==20.7
openpyxl==3.1.5
